+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - TwitchPlugin: Add events TwitchUserSubscribed (both sub and resub), TwitchGiftedSubscription and TwitchRaided
   - TwitchPlugin: Made it Throttle-aware, hopefully avoiding getting locked out of chat for 30mins
   - IRacingPlugin: Added IRacingCarPosition
+  - New plugin: DiscordPlugin
 
 ## [0.4.0](https://github.com/dennis/slipstream/releases/tag/v0.3.0) (2020-01-10)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.3.0...v0.4.0)